Updating the version to 2.0.0
This is a much smaller change than it sounds, but was done because of
some incorrect uses of semver.

This is mostly the same release as 1.2.1, except that its version
of couch2pg has been upgraded to fix a unicode parsing bug.

The 2.0.0 release branch is a legacy branch that does not support
multiple CouchDBs databases. All future work, unless there are critical
support issues, will be done in a 3.x.x release (which does use a
version of couch2pg that supports multiple CouchDB databases).