Upgrading the mainline to 3.0.0
This release is identical to 1.4.0. This is a rejigging of semvers to
accomodate older releases getting patch updates.

The 2.x.x line is for the 2.x.x version of medic/medic, where you only
pull from one CouchDB database.

The 3.x.x line is for the 3.x.x version of medic/medic, where you pull
from multiple.