Move readme to docs (#150)
This compliments https://github.com/medic/cht-docs/pull/1807 and removes
all of readme in favor of the soon to be created item in CHT Docs.

Note - until the [other PR](https://github.com/medic/cht-docs/pull/1807)
is merged [the target
URL](https://docs.communityhealthtoolkit.org/hosting/couch2pg/setup-and-devlopment/)
will `404`. This should work in the next day or so!